omxh264enc: Localize ref-frames property
Specification of nRefFrames of RZ is different from
Khronos specification, and 0 can be specified for RZ.

Actual limit for nRefFrames for RZ is from 0 to 2,
but set G_MAXUINT for upper limit for backward compatibility.
If specify larger than 2, component automatically modify
according to other parameters.

Change identify RCAR to RZ
Signed-off-by: trungvanle <trung.le.xk@renesas.com>