gstomxvideodec: gst-pipeline cannot corectly decode with vspmfilter and filesink/fakesink
The error conditions are that:
	1. Videosink is fakesink or filesink
	2. Using vspmfilter in dmabuf-use, or outbuf-alloc mode
	3. Do not scale
	4. Output video format has 2 or 3 planes

In above conditions, the API "GST_VIDEO_META_API_TYPE" is not added to query from downstream.
This means that omxh264dec will not set "GST_BUFFER_POOL_OPTION_VIDEO_META" config to
buffer pool when allocating output buffers. As a result, the buffers are sent to
vspmfilter from omxh264dec which maps by buffermap instead of framemap. Hence, if the
video format has 2 or 3 planes, the mapping will be incorrect. Because of wrong address,
it causes 2 difference phenomena on RZG2L and RZG2 series:
	1. RZG2L: Error -412 will be output if use fakesink/filesink
	2. RZG2 series: Output file will be wrong if use filesink

-> Solution: Flag GST_BUFFER_POOL_OPTION_VIDEO_META will be added to the buffer pool as
default to ensure that framemap is always being used to map the buffers instead of buffermap
in case that framemap is required from downstream. Based on meta information, the buffer is mapped
correctly. So the error will be fixed

Signed-off-by: trungvanle <trung.le.xk@renesas.com>