gstomxvideoenc: Support updating stride and sliceheight using input buffer metadata
When using vspmfilter as upstream, the stride of output from vspmfilter may
round up due to ISU limitations. However, omxh264enc retrieves information from
caps and sets frame width as stride to input port that is not updated. This results
in wrong display output.
To solve this issue, we need to update stride and sliceheight based on input buffer
metadata provided from upstream when enabling component.

Signed-off-by: trungvanle <trung.le.xk@renesas.com>

Porting from 1.16.3 to 1.22.11
Signed-off-by: trungvanle <trung.le.xk@renesas.com>