omxvideodec: Override sink_query
gst-omx decoder does not work as same as video filter / converter. The
difference of resolution between sink pad and source pad is not allowed.

This patch is a work-around to help gst-omx decoder pass the negotiation phase
if it uses scaling feature. It will return the template caps of sink pad
to upstream instead of a proxy caps.

Signed-off-by: Kiet Pham <kiet.pham.xb@renesas.com>

Porting from 1.16.3 to 1.22.11
Signed-off-by: trungvanle <trung.le.xk@renesas.com>