omxvideodec: Modify the conversion of ARGB8888 and ARGB8888 from OMX to GST format
There is a mismatch between OMX IL Spec 1.1.2 and description of
OMX_COLOR_Format32bitARGB8888 and OMX_COLOR_Format32bitBGRA8888. This can cause
confusion when developing and using these formats. Their description are also
fixed in OMX IL Spec 1.2.0.

So, we decided to convert the 32-bit RGB formats based on the OMX enumerate
instead of the spec description.
  - OMX_COLOR_Format32bitARGB8888 --> ARGB.
  - OMX_COLOR_Format32bitBGRA8888 --> BGRA.

Signed-off-by: Kiet Pham <kiet.pham.xb@renesas.com>